drm/rockchip: ebc: Continue driving sdclk at the end of the data line
Some EINK panels require dummy clocks to be inserted at the end of the
data line. This means ebc need to keep driving the sdclk during LEL.

Some EINK panels do not specify whether dummy clocks need to be
inserted, but the test results show that there is no effect on
the display regardless of whether they are inserted.

Signed-off-by: Chaoyi Chen <chaoyi.chen@rock-chips.com>
Change-Id: Id7d00882548836356abaa18ef98178163fc50651